Move Format to pkg/model
