+++ pkg/model/types.go
@@ -15,3 +15,13 @@ type Repository struct {
 	Source string `json:"source" bson:"source"`
 	ID     string `json:"id,omitempty" bson:"id,omitempty"`
 }
+
+// Format represents the input format type
+type Format string
+
+const (
+	FormatString   Format = "string"
+	FormatNumber   Format = "number"
+	FormatBoolean  Format = "boolean"
+	FormatFilePath Format = "file_path"
+)
